direction change, keeping notes
to trigger Angular's change-detection, I need to pass an Input property to meteorite-list.
Will create a parent component that has the filter meteorites div and button, meteorite-list will be nested in the template as it's child.
the main page parent component will have filterBy property, and set its default value, meteorite-list will take this in as an Input property.
The button click will change the property value, and the change will be detected and trigger ngOnChanges in the child component, meteorite-list