functional - but only without caching
initial though was correct, with this setup the cache is getting in the way.
The reason is that with the same get request endpoint, the application returns the cache.
Specifically for this instance, I need to not return the cached response.
so this get request endpoint needs to be left out of the cache.
But, when I save this to an array, this should work.
Given I'm only console.logging directly from the service, I am able to get a new http request each time the button is pressed.
I could put this functionality in the component instead of the directive perhaps