ngIf statement for non-image responses, functioning well
Going to start to integrate this now.
I want this to be in the astro-pics thumbnail page at the bottom
I'll need to have the two-grid view for the main section, and the bottom will look something like the first fold of the home page, but with a different image
I'm thinking a UFO svg would be nice there
Should be fairly straightforward to copy that
What I'll need to do is create a package, or make the breakpoints service reusable
Would be awesome to create and publish an npm package!